add new API for making hooks on the server $thing->hook['item.create']->create()
To easily and quickly create a hook, use a PodioApp, PodioApp/Field, or
PodioWorkspace object with the hook variable, which is a virtual array
indexed by hook type.  The create() or remove() method is used to add
or remove hooks.
This is fully tested now for apps and app fields, workspace is next