fix iterating over references
foreach ($item->fields['reference'] as $ref) {
    $ref->fields['blah'] = 'hi';
    $ref->save();
}

until this commit, the code above had a subtle logic bug, where $ref
was a Chiara\PodioItem\Values\App object, and since this extends
Chiara\PodioItem\Field, the save() method was actually called field
save instead of app save.

Now the save() method works as one expects.

In addition, the logic was not preserving any changes made to the
object, this fixes that issue as well.