finish filters.  docs and tests
This is working and documented now.

There are some unanswered questions about the format
of what approved/rsvp values should look like, and how
to filter by items with no ratings.  In addition, confirmation
of how created_via works is needed (I am on an airplane, and not paying
for wifi just for that).  Otherwise this is
good to go.